+++ pytket/docs/changelog.md
@@ -3,7 +3,7 @@
 # 2.17.0 (unreleased)
 
 Fixes:
-- Fix handling of wasm function calls in circuit boxes 
+- Fix handling of wasm function calls in circuit boxes
 
 ## 2.16.0 (March 2026)